refactor: many refactors

+++ src/app/controllers/ContactController.js
@@ -1,15 +1,22 @@
 const ContactsRepository = require('../repositories/ContactsRepository');
+const isValidUUID = require('../../utils/isValidUUID');
 
 class ContactController {
   async index(request, response) {
     const { orderBy } = request.query;
     const contacts = await ContactsRepository.findAll(orderBy);
 
-    response.json(contacts);
+    return response.json(contacts);
   }
 
   async show(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({
+        error: 'Invalid contact id',
+      });
+    }
 
     const contact = await ContactsRepository.findById(id);
 
@@ -17,7 +24,7 @@ class ContactController {
       return response.status(404).json({ error: 'Contact not found' });
     }
 
-    response.json(contact);
+    return response.json(contact);
   }
 
   async store(request, response) {
@@ -25,56 +32,89 @@ class ContactController {
       name, email, phone, category_id,
     } = request.body;
 
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({
+        error: 'Invalid category',
+      });
+    }
+
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactExists = await ContactsRepository.findByEmail(email);
+    if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
 
-    if (contactExists) {
-      return response.status(400).json({ error: 'This e-mail is already in use' });
+      if (contactExists) {
+        return response.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.create({
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
 
-    response.status(201).json(contact);
+    return response.status(201).json(contact);
   }
 
   async update(request, response) {
     const { id } = request.params;
-
     const {
       name, email, phone, category_id,
     } = request.body;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({
+        error: 'Invalid contact id',
+      });
+    }
+
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({
+        error: 'Invalid category',
+      });
+    }
+
+    if (!name) {
+      return response.status(400).json({ error: 'Name is required' });
+    }
 
     const contactExists = await ContactsRepository.findById(id);
     if (!contactExists) {
       return response.status(404).json({ error: 'Contact not found' });
     }
 
-    if (!name) {
-      return response.status(400).json({ error: 'Name is required' });
-    }
-
-    const contactByEmail = await ContactsRepository.findByEmail(email);
-    if (contactByEmail && contactByEmail.id !== id) {
-      return response.status(400).json({ error: 'This e-mail is already in use' });
+    if (email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        return response.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.update(id, {
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
 
-    response.json(contact);
+    return response.json(contact);
   }
 
   async delete(request, response) {
     const { id } = request.params;
 
+    if (!isValidUUID(id)) {
+      return response.status(400).json({
+        error: 'Invalid contact id',
+      });
+    }
+
     await ContactsRepository.delete(id);
-    response.sendStatus(204);
+    return response.sendStatus(204);
   }
 }
